refactor: rename package
rename package from com.ssafy.todo to com.ssafy.sandbox
