+++ todo/src/main/java/com/ssafy/sandbox/TodoApplication.java
@@ -1,4 +1,4 @@
-package com.ssafy.todo;
+package com.ssafy.sandbox;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
