+++ todo/src/main/java/com/ssafy/sandbox/model/repository/TodoRepository.java
@@ -1,7 +1,7 @@
-package com.ssafy.todo.model.repository;
+package com.ssafy.sandbox.model.repository;
 
-import com.ssafy.todo.model.domain.Todo;
-import com.ssafy.todo.model.domain.TodoVO;
+import com.ssafy.sandbox.model.domain.Todo;
+import com.ssafy.sandbox.model.domain.TodoVO;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
