+++ todo/src/test/java/com/ssafy/sandbox/TodoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.ssafy.todo;
+package com.ssafy.sandbox;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
